app/requirements.txt
amqp==2.5.2
asgiref==3.2.5
beautifulsoup4==4.8.2
billiard==3.6.3.0
biopython==1.76
celery==4.4.2
celery-progress==0.0.9
certifi==2019.11.28
chardet==3.0.4
construct==2.5.3
Django==3.0.4
django-bootstrap4==1.1.1
django-cors-headers==3.2.1
django-haystack==3.0b1
future==0.18.2
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
kombu==4.6.8
numpy==1.18.2
pandas==1.0.3
pathtools==0.1.2
pefile==2019.4.18
pubmed-lookup==0.2.3
pysolr==3.8.1
python-dateutil==2.8.1
python-ptrace==0.9.4
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
tqdm==4.43.0
urllib3==1.25.8
vine==1.3.0
xmltodict==0.12.0
zipp==3.1.0
